Refactor control movement into its own module and apply handlers to the top-level canvas
In the future, we'll want to support multi-control movement/resizing, and this will make it easier.  Plus the code is more straight forward

Also adds back undo/redo support for movement; completing item "Undo/Redo" of #33
